brain/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2
psycopg2-binary>=2.9
pgvector>=0.2
faiss-cpu>=1.7
numpy>=1.24
sentence-transformers>=2.2
torch>=2.0
transformers>=4.30
requests>=2.31
python-multipart>=0.0.6
aiofiles>=23.0
python-dotenv>=1.0
zep-python>=2.0
openai>=1.15